+++ demo-app/changelog.md
@@ -18,7 +18,8 @@ The `svg-dom-demo` gallery is not published as part of the crate's release, so t
 - Correct effect of spot slider on beam (`16b6bd4`)
 - Correct overstatement in LightSource documentation and adjust demo (`26c80d5`)
 - Doc only: update stale doc comment (`458982d`)
-- Cone test verifies 0º Chrome anomoly (``) 
+- Cone test verifies 0º Chrome anomoly (`d43d3b8`)
+- Improve test coverage (``)
 
 ## [0.1.15] - 2026-08-06